+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### NEXT
+
+* Sync ortc with mediasoup-client ([#198](https://github.com/versatica/libmediasoupclient/pull/198)).
 
 ### 3.5.0